requirements.txt
# ============================================================
# Core Dependencies
# ============================================================

# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph-checkpoint-sqlite>=1.0.0

# OpenAI
openai>=1.50.0

# ============================================================
# Web Frameworks
# ============================================================

# FastAPI for REST API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Streamlit for GUI
streamlit>=1.40.0

# ============================================================
# API Integrations
# ============================================================

# HTTP client for external APIs
httpx>=0.27.0
aiohttp>=3.10.0

# ============================================================
# Authentication & Security
# ============================================================

# JWT authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ============================================================
# Database
# ============================================================

# SQLite (included in Python)
# For production, add:
# sqlalchemy>=2.0.0
# asyncpg>=0.29.0  # PostgreSQL
# redis>=5.0.0     # Caching

# ============================================================
# Data Validation
# ============================================================

pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0

# ============================================================
# Development & Testing
# ============================================================

pytest>=8.3.0
pytest-cov>=5.0.0
pytest-asyncio>=0.24.0

# Linting & Formatting
ruff>=0.7.0
black>=24.10.0
isort>=5.13.0
mypy>=1.12.0

# Security scanning
safety>=3.2.0
bandit>=1.7.0

# ============================================================
# Utilities
# ============================================================

# Rich console output
rich>=13.9.0

# Date/time handling
python-dateutil>=2.9.0

# Typing extensions
typing-extensions>=4.12.0
